+++ packages/http-client-csharp/generator/Microsoft.TypeSpec.Generator/src/Providers/TypeProvider.cs
@@ -300,7 +300,7 @@ namespace Microsoft.TypeSpec.Generator.Providers
                 }
             }
 
-            return [..properties];
+            return [.. properties];
         }
 
         internal FieldProvider[] FilterCustomizedFields(IEnumerable<FieldProvider> specFields)
@@ -339,7 +339,7 @@ namespace Microsoft.TypeSpec.Generator.Providers
                 }
             }
 
-            return [..methods];
+            return [.. methods];
         }
 
         internal ConstructorProvider[] FilterCustomizedConstructors(IEnumerable<ConstructorProvider> specConstructors)
@@ -353,7 +353,7 @@ namespace Microsoft.TypeSpec.Generator.Providers
                 }
             }
 
-            return [..constructors];
+            return [.. constructors];
         }
 
         private TypeProvider[] BuildNestedTypesInternal()
@@ -507,27 +507,36 @@ namespace Microsoft.TypeSpec.Generator.Providers
 
             if (name != null)
             {
-                // Reset the custom code view to reflect the new name
-                _customCodeView = new(BuildCustomCodeView(name, Type.Namespace));
-                // Give precedence to the custom code view name if it exists
-                _lastContractView = new(BuildLastContractView(
-                    _customCodeView.Value?.Name ?? name,
-                    _customCodeView.Value?.Type.Namespace ?? Type.Namespace));
-                Type.Update(name: _customCodeView.Value?.Name ?? name, @namespace: _customCodeView.Value?.Type.Namespace);
+                ResetMembersBasedOnIdentityChange(name);
             }
 
             if (@namespace != null)
             {
-                // Reset the custom code view to reflect the new namespace
-                _customCodeView = new(BuildCustomCodeView(Type.Name, @namespace));
-                _lastContractView = new(BuildLastContractView(Type.Name, @namespace));
-                _declarationModifiers = BuildDeclarationModifiersInternal(); // recalculate declaration modifiers
-                Type.Update(@namespace: _customCodeView.Value?.Type.Namespace ?? @namespace);
+                ResetMembersBasedOnIdentityChange(@namespace: @namespace);
             }
 
             // Rebuild the canonical view
             _canonicalView = new(BuildCanonicalView);
         }
+
+        private void ResetMembersBasedOnIdentityChange(string? name = null, string? @namespace = null)
+        {
+            // Reset the custom code view to reflect the new namespace
+            _customCodeView = new(BuildCustomCodeView(name ?? Type.Name, @namespace ?? Type.Namespace));
+            name = _customCodeView.Value?.Name ?? name ?? Type.Name;
+            @namespace = _customCodeView.Value?.Type.Namespace ?? @namespace ?? Type.Namespace;
+            _lastContractView = new(BuildLastContractView(
+                name,
+                @namespace));
+            // recalculate declaration modifiers and constructors
+            _declarationModifiers = null;
+            // constructors might change based on declaration modifier changes
+            _constructors = null;
+            // serialization providers need to reflect the new type name/namespace
+            _serializationProviders = null;
+            Type.Update(name: name, @namespace: @namespace);
+        }
+
         public IReadOnlyList<EnumTypeMember> EnumValues => _enumValues ??= BuildEnumValues();
 
         protected virtual IReadOnlyList<EnumTypeMember> BuildEnumValues() => throw new InvalidOperationException("Not an EnumProvider type");
@@ -553,6 +562,19 @@ namespace Microsoft.TypeSpec.Generator.Providers
                 type.EnsureBuilt();
             }
         }
+
+        internal void ProcessTypeForBackCompatibility()
+        {
+            if (LastContractView?.Methods == null || LastContractView?.Methods.Count == 0)
+            {
+                return;
+            }
+
+            Update(methods: BuildMethodsForBackCompatibility(Methods));
+        }
+
+        protected internal virtual IReadOnlyList<MethodProvider> BuildMethodsForBackCompatibility(IEnumerable<MethodProvider> originalMethods)
+            => [.. originalMethods];
 
         private IReadOnlyList<EnumTypeMember>? _enumValues;
 
@@ -666,7 +688,7 @@ namespace Microsoft.TypeSpec.Generator.Providers
             }
             else if (attribute.ConstructorArguments[1].Kind != TypedConstantKind.Array)
             {
-                parameterTypes = attribute.ConstructorArguments[1..].Select(a => (ISymbol?) a.Value).ToArray();
+                parameterTypes = attribute.ConstructorArguments[1..].Select(a => (ISymbol?)a.Value).ToArray();
             }
             else
             {