+++ packages/http-client-java/generator/http-client-generator-core/src/main/java/com/microsoft/typespec/http/client/generator/core/template/StreamSerializationModelTemplate.java
@@ -15,6 +15,7 @@ import com.microsoft.typespec.http.client.generator.core.model.clientmodel.Clien
 import com.microsoft.typespec.http.client.generator.core.model.clientmodel.ClientModelProperty;
 import com.microsoft.typespec.http.client.generator.core.model.clientmodel.ClientModelPropertyAccess;
 import com.microsoft.typespec.http.client.generator.core.model.clientmodel.ClientModelPropertyReference;
+import com.microsoft.typespec.http.client.generator.core.model.clientmodel.ConvertToJsonTypeTrait;
 import com.microsoft.typespec.http.client.generator.core.model.clientmodel.IType;
 import com.microsoft.typespec.http.client.generator.core.model.clientmodel.IterableType;
 import com.microsoft.typespec.http.client.generator.core.model.clientmodel.MapType;
@@ -29,10 +30,12 @@ import com.microsoft.typespec.http.client.generator.core.util.ClientModelUtil;
 import io.clientcore.core.utils.CoreUtils;
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Base64;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
+import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -93,6 +96,9 @@ public class StreamSerializationModelTemplate extends ModelTemplate {
         imports.add(Base64.class.getName());
         imports.add(LinkedHashMap.class.getName());
         imports.add(List.class.getName());
+        imports.add(LinkedList.class.getName());
+        imports.add(Arrays.class.getName());
+        imports.add(Collectors.class.getName());
         imports.add(Map.class.getName());
         imports.add(Objects.class.getName());
     }
@@ -742,9 +748,39 @@ public class StreamSerializationModelTemplate extends ModelTemplate {
                         .line("jsonWriter.writeUntypedField(\"" + serializedName + "\", " + propertyValueGetter + ");");
                 }
             } else if (wireType instanceof IterableType) {
-                serializeJsonContainerProperty(methodBlock, "writeArrayField", wireType,
-                    ((IterableType) wireType).getElementType(), serializedName, propertyValueGetter, 0,
-                    isJsonMergePatch);
+                if (property.getArrayEncoding() == null) {
+                    serializeJsonContainerProperty(methodBlock, "writeArrayField", wireType,
+                        ((IterableType) wireType).getElementType(), serializedName, propertyValueGetter, 0,
+                        isJsonMergePatch);
+                } else {
+                    IType wireElementType = ((IterableType) wireType).getElementType();
+                    if (wireElementType == ClassType.STRING) {
+                        // wireType is String
+                        methodBlock.ifBlock(propertyValueGetter + " != null", ifBlock -> {
+                            String serializeExpression = propertyValueGetter
+                                + ".stream().map(element -> element == null ? \"\" : element).collect(Collectors.joining(\""
+                                + property.getArrayEncoding().getDelimiter() + "\"))";
+                            methodBlock.line("jsonWriter.writeStringField(\"%s\", %s);", serializedName,
+                                serializeExpression);
+                        });
+                    } else {
+                        // wireType need to be converted to String
+                        if (wireElementType instanceof ConvertToJsonTypeTrait) {
+                            String convertToJsonExpression
+                                = ((ConvertToJsonTypeTrait) wireElementType).convertToJsonType("element");
+                            methodBlock.ifBlock(propertyValueGetter + " != null", ifBlock -> {
+                                String serializeExpression = propertyValueGetter + ".stream().map(element -> "
+                                    + convertToJsonExpression + ").collect(Collectors.joining(\""
+                                    + property.getArrayEncoding().getDelimiter() + "\"))";
+                                methodBlock.line("jsonWriter.writeStringField(\"%s\", %s);", serializedName,
+                                    serializeExpression);
+                            });
+                        } else {
+                            throw new RuntimeException("Unable to convert type " + wireElementType
+                                + " to String for ArrayEncoding serialization.");
+                        }
+                    }
+                }
             } else if (wireType instanceof MapType) {
                 // Assumption is that the key type for the Map is a String. This may not always hold true and when that
                 // becomes reality this will need to be reworked to handle that case.
@@ -1587,13 +1623,40 @@ public class StreamSerializationModelTemplate extends ModelTemplate {
                     deserializationBlock.line(property.getName() + " = reader.readUntyped();");
                 }
             } else if (wireType instanceof IterableType) {
+                final String propertyStringVariableName = property.getName() + "EncodedAsString";
+                IType wireElementType = ((IterableType) wireType).getElementType();
+                if (property.getArrayEncoding() != null) {
+                    // need to prepare the expression for propertyStringVariableName,
+                    // to be used in "if (property.getArrayEncoding() == null)" block
+                    deserializationBlock.line("String " + propertyStringVariableName + " = reader.getString();");
+                }
+
                 if (!propertiesManager.hasConstructorArguments()) {
                     deserializationBlock.text(property.getClientType() + " ");
                 }
 
                 deserializationBlock.text(property.getName() + " = ");
-                deserializeJsonContainerProperty(deserializationBlock, "readArray", wireType,
-                    ((IterableType) wireType).getElementType(), ((IterableType) clientType).getElementType(), 0);
+                if (property.getArrayEncoding() == null) {
+                    deserializeJsonContainerProperty(deserializationBlock, "readArray", wireType, wireElementType,
+                        ((IterableType) clientType).getElementType(), 0);
+                } else {
+                    // LinkedList is used to be consistent with internal code of core, e.g. "readArray" API
+                    if (wireElementType == ClassType.STRING) {
+                        // wireType is String
+                        deserializationBlock.line(
+                            "%1$s == null ? null : %1$s.isEmpty() ? new LinkedList<>() : new LinkedList<>(Arrays.asList(%1$s.split(\"%2$s\", -1)));",
+                            propertyStringVariableName, property.getArrayEncoding().getEscapedDelimiter());
+                    } else {
+                        // wireType need to be converted from String
+                        String conversionExpress = wireElementType.defaultValueExpression("valueAsString")
+                            .replace("\"valueAsString\"", "valueAsString");
+                        deserializationBlock.line(
+                            "%1$s == null ? null : %1$s.isEmpty() ? new LinkedList<>() : new LinkedList<>(Arrays.stream(%1$s.split(\"%2$s\", -1)).map(valueAsString -> %3$s).collect(Collectors.toList()));",
+                            propertyStringVariableName, property.getArrayEncoding().getEscapedDelimiter(),
+                            conversionExpress);
+
+                    }
+                }
 
                 if (!propertiesManager.hasConstructorArguments()) {
                     handleSettingDeserializedValue(deserializationBlock, property, property.getName(), fromSuper);