+++ packages/spec-coverage-sdk/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log - @typespec/spec-coverage-sdk
 
+## 0.1.0-alpha.14
+
+### Bump dependencies
+
+- [#9223](https://github.com/microsoft/typespec/pull/9223) Upgrade dependencies
+
+
 ## 0.1.0-alpha.13
 
 ### Bump dependencies